+++ data/examples/declaration/value/function/negative-literals-out.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE NegativeLiterals #-}
+
+foo = - 1
+
+foo = -1
+
+foo = -x
+
+foo = -x