+++ src/Ormolu/Printer/Meat/Declaration/Value.hs
@@ -32,7 +32,7 @@ import GHC.Hs.Extension
 import GHC.Hs.Lit
 import GHC.Hs.Pat
 import GHC.Hs.Type
-import GHC.LanguageExtensions.Type (Extension (LexicalNegation))
+import GHC.LanguageExtensions.Type (Extension (NegativeLiterals))
 import GHC.Parser.Annotation
 import GHC.Parser.CharClass (is_space)
 import GHC.Types.Basic
@@ -621,9 +621,15 @@ p_hsExpr' s = \case
     let opTree = OpBranch (exprOpTree x) op (exprOpTree y)
     p_exprOpTree s (reassociateOpTree getOpName opTree)
   NegApp NoExtField e NoExtField -> do
-    lexicalNegation <- isExtensionEnabled LexicalNegation
+    negativeLiterals <- isExtensionEnabled NegativeLiterals
+    let isLiteral = case unLoc e of
+          HsLit {} -> True
+          HsOverLit {} -> True
+          _ -> False
     txt "-"
-    unless lexicalNegation space
+    -- If NegativeLiterals is enabled, we have to insert a space before
+    -- negated literals, as `- 1` and `-1` have differing AST.
+    when (negativeLiterals && isLiteral) space
     located e p_hsExpr
   HsPar NoExtField e ->
     parens s (located e (dontUseBraces . p_hsExpr))